+++ glue/viewers/histogram/qt/tests/test_data_viewer.py
@@ -511,7 +511,7 @@ class TestHistogramViewer(object):
         # messages regarding subset updates (this occurred when multiple)
         # datasets were present.
 
-        layer_tree = LayerTreeWidget()
+        layer_tree = LayerTreeWidget(session=self.session)
         layer_tree.set_checkable(False)
         layer_tree.setup(self.data_collection)
         layer_tree.bind_selection_to_edit_subset()